requirements.txt
# Flask and Extensions
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
Flask-Bcrypt>=1.0.1
Flask-CORS>=4.0.0

# Database
PyMySQL>=1.1.0
cryptography>=41.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Production Server
gunicorn>=21.0.0

# Development
pytest>=7.4.0
pytest-flask>=1.3.0
